requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
aiohttp>=3.9.0
asyncio>=3.4.3

# Browser automation (optional)
playwright>=1.40.0
selenium>=4.15.0
webdriver-manager>=4.0.0

# CLI & Progress
click>=8.1.0
tqdm>=4.66.0
rich>=13.7.0

# Data handling
pandas>=2.1.0
python-dotenv>=1.0.0

# Rate limiting & retries
ratelimit>=2.2.1
tenacity>=8.2.0
fake-useragent>=1.4.0

# Logging
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
responses>=0.24.0

# Utilities
urllib3>=2.1.0
certifi>=2023.11.0
